edit README.md

+++ main.py
@@ -132,9 +132,10 @@ from src.decorators import log
 # for card_number in card_number_generator(8952458625865240, 8952458625865245):
 #     print(card_number)
 
-@log('mylog.txt')
+
+@log("mylog.txt")
 # @log()
-def my_function(x: int, y: int) -> int:
+def my_function(x: int, y: int) -> float:
     return x / y
